Rename lib to inferno and organize into modules
